\name{exdata}
\alias{exdata}
\docType{data}
\title{
 Example Data
}
\description{
Data generated by dynamic linear model including 2 target series (one with linear trend and seasonality; another with generalized trend and cycle) and 16 contemporaneous predictors. 
}
\usage{data(exdata)}
\format{
A data frame with 500 observations on the following 18 variables. First two variables are 2 target series. The next 8 variables are candidate predictors for 1st target series, while the rest of them are candidate predictors for 2nd target series. 
}

\source{
See Equation (40) in the paper worte by Jamalamadaka, Qiu and Ning (2018) <https://arxiv.org/pdf/1801.03222.pdf>
}

\examples{
data(exdata)
ts.plot(exdata[,1:2])
}
\keyword{datasets}
