% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mbsts-package}
\alias{mbsts-package}
\title{Multivariate Bayesian Structural Time Series}
\description{
Tools for data analysis with multivariate Bayesian structural time series (MBSTS) models.  Specifically, the package provides facilities for implementing general structural time series models, flexibly adding on different time series components (trend, season, cycle, and regression), simulating them, fitting them to multivariate correlated time series data, conducting feature selection on the regression component.
}
\section{Documentation}{

\pkg{mbsts} is described in Ning and Qiu (2021).
}

\section{License}{

\pkg{mbsts} is provided under the \acronym{LGPL-2.1} License.
}

\references{
\Qiu2018

\Jammalamadaka2019

\Ning2021
}
\author{
Jinwen Qiu, Ning Ning
}
