% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Simulated Example Data}
\format{
A tibble with 505 rows and 18 variables:
}
\usage{
simdata
}
\description{
Data generated by a dynamic linear model including two target series (one with a linear trend component and a seasonal component; the other with a generalized trend component and a cycle component) and \eqn{16} contemporaneous predictors.
}
\examples{
ts.plot(simdata[,1:2],col = 1:ncol(simdata[,1:2]))
legend("topleft", colnames(simdata[,1:2]), col=1:ncol(simdata[,1:2]), lty=1, cex=.65)
}
\references{
\Ning2021
}
\author{
Jinwen Qiu \email{qjwsnow_ctw@hotmail.com} Ning Ning \email{patricianing@gmail.com}
}
\keyword{datasets}
